move *Loader to grails.geb.serviceloader. as per https://github.com/grails/geb/pull/146#issuecomment-2686107413
intellij refactor + Spec needed manually
